Refactor `LLM Query Options` menu: extract the options dialog to new file -- CW-710
