+++ src/features/LLMQuery/components/index.tsx
@@ -1,5 +1,6 @@
+import { LLMQueryOptionsDialog } from './LLMQueryOptionsDialog'
 import { LLMQueryOptionsMenuItem } from './LLMQueryOptionsMenuItem'
 import { LLMQueryResultPanel } from './LLMQueryResultPanel'
 import { RunLLMQueryMenuItem } from './RunLLMQueryMenuItem'
 
-export { LLMQueryOptionsMenuItem, LLMQueryResultPanel, RunLLMQueryMenuItem }
+export { LLMQueryOptionsDialog, LLMQueryOptionsMenuItem, LLMQueryResultPanel, RunLLMQueryMenuItem }
