+++ src/features/ToolBar/AnalysisMenu/index.tsx
@@ -2,6 +2,7 @@ import { MenuItem } from 'primereact/menuitem'
 import { useState } from 'react'
 
 import {
+  LLMQueryOptionsDialog,
   LLMQueryOptionsMenuItem,
   RunLLMQueryMenuItem,
 } from '../../LLMQuery/components'
@@ -10,9 +11,19 @@ import { DropdownMenu } from '../DropdownMenu'
 
 export const AnalysisMenu = () => {
   const [open, setOpen] = useState(false)
+  const [openDialog, setOpenDialog] = useState(false)
 
   const handleClose = (): void => {
     setOpen(false)
+  }
+
+  const handleOpenDialog = (): void => {
+    handleClose()
+    setOpenDialog(true)
+  }
+
+  const handleCloseDialog = (): void => {
+    setOpenDialog(false)
   }
 
   const menuItems: MenuItem[] = [
@@ -23,17 +34,23 @@ export const AnalysisMenu = () => {
       separator: true,
     },
     {
-      template: <LLMQueryOptionsMenuItem onClick={handleClose} />,
+      template: <LLMQueryOptionsMenuItem onClick={handleOpenDialog} />,
     },
   ]
 
   return (
-    <DropdownMenu
-      id="analysis-menu"
-      label="Analysis"
-      menuItems={menuItems}
-      open={open}
-      onOpenChange={setOpen}
-    />
+    <>
+      <DropdownMenu
+        id="analysis-menu"
+        label="Analysis"
+        menuItems={menuItems}
+        open={open}
+        onOpenChange={setOpen}
+      />
+      <LLMQueryOptionsDialog
+        open={openDialog}
+        handleClose={handleCloseDialog}
+      />
+    </>
   )
 }
